+++ src/messager.rs
@@ -1,39 +1,35 @@
-use colored::Colorize;
-use log::{error, info, warn};
 
-use crate::{messager, ncmdump,AppError};
+
+use crate::{messager, AppError};
 use std::fmt::Debug;
-use std::sync::mpsc;
+// use
 pub struct Messager {
     name: String,
-    sender: mpsc::Sender<messager::Message>,
+    sender: crossbeam_channel::Sender<messager::Message>,
 }
 
 pub struct Message {
     pub name: String,
     pub signal: Signals,
 }
-impl Message {
-    // 定义一个公共方法 log，用于记录不同信号状态下的日志信息
-    pub fn log(&self) {
-        let loginfo = match &self.signal {
-            Signals::Start => "读取文件",
-            Signals::GetMetaInfo => "解密歌曲元信息",
-            Signals::GetCover => "解密封面图片数据",
-            Signals::Decrypt => "解密歌曲信息",
-            Signals::Save => "保存文件",
-            Signals::End => "成功!",
-            Signals::Err(e) => &e.to_string(),
-        };
-        match &self.signal {
-            Signals::Err(e) => match e {
-                AppError::ProtectFile => warn!("[{}] {}", self.name.cyan(), loginfo),
-                _ => error!("[{}] {}", self.name.cyan(), loginfo),
-            },
-            _ => info!("[{}] {}", self.name.cyan(), loginfo),
-        }
-    }
-}
+// impl Message {
+//     // 定义一个公共方法 log，用于记录不同信号状态下的日志信息
+//     pub fn log(&self) {
+
+//         match &self.signal {
+//             Signals::Err(e) => match e {
+//                 AppError::ProtectFile => warn!("[{}] {}", self.name.cyan(), "强制覆盖已关闭。不保存文件"),
+//                 _ => error!("[{}] {}", self.name.cyan(), e),
+//             },
+//             Signals::Start => trace!("开始读取文件"),
+//             Signals::GetMetaInfo => trace!("解密歌曲元信息"),
+//             Signals::GetCover => trace!("解密封面图片数据"),
+//             Signals::Decrypt => trace!("解密歌曲信息"),
+//             Signals::Save => info!("保存文件"),
+//             Signals::End => trace!("解密流程结束")
+//         }
+//     }
+// }
 #[derive(PartialEq)]
 pub enum Signals {
     Start,
@@ -46,10 +42,10 @@ pub enum Signals {
 }
 
 impl Messager {
-    pub fn new(name: String, sender: mpsc::Sender<messager::Message>) -> Self {
+    pub fn new(name: String, sender: crossbeam_channel::Sender<messager::Message>) -> Self {
         Self { name, sender }
     }
-    pub fn send(&self, s: Signals) -> Result<(), std::sync::mpsc::SendError<messager::Message>> {
+    pub fn send(&self, s: Signals) -> Result<(), crossbeam_channel::SendError<messager::Message>> {
         self.sender.send(Message {
             name: self.name.clone(),
             signal: s,